refactor: Update suggestion changes
